% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2points-package.R
\docType{package}
\name{raster2points-package}
\alias{raster2points-package}
\title{raster2points: Raster2points}
\description{
This function takes a raster file as input and generates a shapefile with points corresponding to the raster's pixel values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Rene-Jubile/raster2points}
  \item Report bugs at \url{https://github.com/Rene-Jubile/raster2points/issues}
}

}
\author{
\strong{Maintainer}: MABILA Rene-Jubile \email{jubile@mabilarene.tech} (\href{https://orcid.org/0000-0000-0000-0000}{ORCID})

}
\keyword{internal}
