% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster2points.R
\name{raster2points}
\alias{raster2points}
\title{This function takes a raster file as input and generates a shapefile with points corresponding to the raster's pixel values.}
\usage{
raster2points(raster_path, shapefile_path)
}
\arguments{
\item{raster_path}{A character string specifying the path to the input raster file.}

\item{shapefile_path}{A character string specifying the path to save the output shapefile.}
}
\description{
This function takes a raster file as input and generates a shapefile with points corresponding to the raster's pixel values.
}
\examples{
# Créer un raster simple pour l'exemple

library(raster)
library(sf)
library(sp)
raster_path <- tempfile(fileext = ".tif")
r <- raster(nrow = 10, ncol = 10, xmn = 0, xmx = 10, ymn = 0, ymx = 10)
values(r) <- 1:ncell(r)
writeRaster(r, raster_path, format = "GTiff", overwrite = TRUE)

# Définir le chemin pour le shapefile de sortie
shapefile_path <- tempfile(fileext = ".shp")

# Utiliser la fonction pour créer le shapefile à partir du raster
raster2points(raster_path, shapefile_path)

# Charger le shapefile créé pour vérifier son contenu
shapefile_points <- st_read(shapefile_path)

# Afficher les premières lignes du shapefile pour vérifier
print(head(shapefile_points))
}
